UPDATE code to be seperated into files
